backend/pyproject.toml
[project]
name = "certification-challenge"
version = "0.1.0"
description = "AIE9 Certification Challenge: AI Assistant for Service Businesses"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-tavily>=0.1.0",
    "langsmith>=0.2.0",
    "langgraph>=0.3.0",
    "qdrant-client>=1.12.0",
    "langchain-qdrant>=0.2.0",
    "python-dotenv>=1.0.1",
    "nest-asyncio>=1.6.0",
    "tiktoken>=0.8.0",
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "ragas==0.2.10",
    "cohere>=5.12.0,<5.13.0", 
    "rapidfuzz>=3.14.3",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
]